Fixed several warnings. CMAKE_COMPILER_IS_CLANGXX was being set too late, causing many flags to be missed if Clang was installed.

+++ OgreMain/src/OgreSubMesh2.cpp
@@ -945,7 +945,7 @@ namespace Ogre {
                     {
                         tangentSrc = &(*itor);
 
-                        assert( (itor - srcData.begin() >= srcData.size() - 2u) &&
+                        assert( ((size_t)(itor - srcData.begin()) >= srcData.size() - 2u) &&
                                 (srcData.back().element.mSemantic == VES_TANGENT ||
                                  srcData.back().element.mSemantic == VES_BINORMAL ) &&
                                 "Tangent element must be at the end of srcData array!" );
@@ -954,7 +954,7 @@ namespace Ogre {
                     {
                         binormalSrc = &(*itor);
 
-                        assert( (itor - srcData.begin() >= srcData.size() - 2u) &&
+                        assert( ((size_t)(itor - srcData.begin()) >= srcData.size() - 2u) &&
                                 (srcData.back().element.mSemantic == VES_TANGENT ||
                                  srcData.back().element.mSemantic == VES_BINORMAL ) &&
                                 "Binormal element must be at the end of srcData array!" );
